Add security.query_api_keys definition
