+++ output/schema/validation-errors.json
@@ -1670,12 +1670,6 @@
       ],
       "response": []
     },
-    "security.query_api_keys": {
-      "request": [
-        "Missing request & response"
-      ],
-      "response": []
-    },
     "security.saml_authenticate": {
       "request": [
         "Missing request & response"
