+++ specification/security/authenticate/SecurityAuthenticateResponse.ts
@@ -17,9 +17,10 @@
  * under the License.
  */
 
+import { ApiKey } from '@security/_types/ApiKey'
 import { RealmInfo } from '@security/_types/RealmInfo'
 import { Metadata, Name, Username } from '@_types/common'
-import { ApiKey, Token } from './types'
+import { Token } from './types'
 
 export class Response {
   body: {
