+++ specification/security/authenticate/types.ts
@@ -24,8 +24,3 @@ export class Token {
   /** @since 7.14.0 */
   type?: string
 }
-
-export class ApiKey {
-  id: string
-  name: Name
-}
