+++ specification/security/grant_api_key/types.ts
@@ -22,7 +22,7 @@ import { UserDefinedValue } from '@spec_utils/UserDefinedValue'
 import { Name } from '@_types/common'
 import { Time } from '@_types/Time'
 
-export class ApiKey {
+export class GrantApiKey {
   name: Name
   expiration?: Time
   role_descriptors?: Dictionary<string, UserDefinedValue>[]
