+++ specification/security/query_api_keys/QueryApiKeysRequest.ts
@@ -1,0 +1,56 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { RequestBase } from '@_types/Base'
+import { integer } from '@_types/Numeric'
+import { QueryContainer } from '@_types/query_dsl/abstractions'
+import { Sort, SortResults } from '@_types/sort'
+
+/**
+ * @rest_spec_name security.query_api_keys
+ * @since 7.15.0
+ * @stability stable
+ */
+export interface Request extends RequestBase {
+  body: {
+    /**
+     * A query to filter which API keys to return.
+     * The query supports a subset of query types, including match_all, bool, term, terms, ids, prefix, wildcard, and range.
+     * You can query all public information associated with an API key
+     */
+    query?: QueryContainer
+    /**
+     * Starting document offset. By default, you cannot page through more than 10,000
+     * hits using the from and size parameters. To page through more hits, use the
+     * search_after parameter.
+     * @server_default 0
+     */
+     from?: integer
+     /** @doc_url https://www.elastic.co/guide/en/elasticsearch/reference/current/sort-search-results.html */
+    sort?: Sort
+    /**
+     * The number of hits to return. By default, you cannot page through more
+     * than 10,000 hits using the from and size parameters. To page through more
+     * hits, use the search_after parameter.
+     * @server_default 10
+     */
+     size?: integer
+     search_after?: SortResults
+  }
+}
